08-flask-snippets/requirements.txt
Flask==2.1.0
psycopg2-binary==2.9.5
requests==2.18.4
python-dotenv==1.0.0
gunicorn==21.0.0
